+++ erpnext/accounts/doctype/pay_period_date/pay_period_date.json
@@ -83,10 +83,10 @@
  "issingle": 0, 
  "istable": 1, 
  "max_attachments": 0, 
- "modified": "2017-07-07 13:31:54.800149", 
+ "modified": "2017-07-12 09:11:57.979040", 
  "modified_by": "Administrator", 
  "module": "Accounts", 
- "name": "Fiscal Year Pay Date", 
+ "name": "Pay Period Date", 
  "name_case": "", 
  "owner": "Administrator", 
  "permissions": [],
