+++ erpnext/accounts/doctype/pay_period_date/pay_period_date.py
@@ -6,5 +6,5 @@ from __future__ import unicode_literals
 import frappe
 from frappe.model.document import Document
 
-class FiscalYearPayDate(Document):
+class PayPeriodDate(Document):
 	pass
